# specs/001-mvp-core-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QualyIT API
  description: API for QualyIT Quality Management Platform (MVP)
  version: 1.0.0
  contact:
    name: QualyIT Support
servers:
  - url: https://{tenant}.qualyit.app/api
    variables:
      tenant:
        default: demo
        description: Tenant subdomain
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TaskStatus'
        - name: areaId
          in: query
          schema:
            type: string
            format: uuid
        - name: assignedToId
          in: query
          schema:
            type: string
            format: uuid
        - name: dueDate
          in: query
          schema:
            type: string
            format: date
          description: Filter by due date
        - name: today
          in: query
          schema:
            type: boolean
          description: Only tasks due today
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskWithDetails'

    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
...
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task details
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task details with checklist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskWithChecklist'

    patch:
      tags: [Tasks]
      summary: Update task
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated

    delete:
      tags: [Tasks]
      summary: Cancel task
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task cancelled
  /tasks/{taskId}/complete:
    post:
      tags: [Tasks]
      summary: Complete task
      description: Mark task as complete (OK or problem)
      parameters:
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteTaskRequest'
      responses:
        '200':
          description: Task completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCompletion'
  /tasks/{taskId}/checklist/{itemId}/complete:
    post:
      tags: [Tasks]
      summary: Complete checklist item
      description: Mark individual checklist item as OK or problem
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - name: itemId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteChecklistItemRequest'
      responses:
        '200':
          description: Item completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TaskType:
      type: string
      enum: [scheduled, corrective, preventive]
    TaskPriority:
      type: string
      enum: [critical, high, medium, low]
    TaskStatus:
      type: string
      enum: [pending, in_progress, completed, cancelled]
    ChecklistStatus:
      type: string
      enum: [pending, ok, problem]
    ProblemReason:
      type: string
      enum: [no_time, no_supplies, equipment_broken, other]

    # ---- Core Entities ----
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        type:
          $ref: '#/components/schemas/TaskType'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        areaId:
          type: string
          format: uuid
        assignedToId:
          type: string
          format: uuid
        dueDate:
          type: string
          format: date-time
        scheduledTime:
          type: string
          format: time
        hasChecklist:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TaskWithDetails:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            areaName:
              type: string
            assignedToName:
              type: string
            checklistProgress:
              type: object
              properties:
                completed:
                  type: integer
                total:
                  type: integer
    TaskWithChecklist:
      allOf:
        - $ref: '#/components/schemas/TaskWithDetails'
        - type: object
          properties:
            checklistItems:
              type: array
              items:
                $ref: '#/components/schemas/ChecklistItem'
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        sortOrder:
          type: integer
        status:
          $ref: '#/components/schemas/ChecklistStatus'
        completedAt:
          type: string
          format: date-time
        completedByName:
          type: string
        problemReason:
          type: string
    TaskCompletion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
        status:
          type: string
          enum: [ok, problem]
        notes:
          type: string
        completedAt:
          type: string
          format: date-time
        files:
          type: array
          items:
            type: string
...
    CreateTaskRequest:
      type: object
      required: [title, areaId, type]
      properties:
        title:
          type: string
          maxLength: 255
        description:
          type: string
        type:
          $ref: '#/components/schemas/TaskType'
        priority:
          $ref: '#/components/schemas/TaskPriority'
        areaId:
          type: string
          format: uuid
        assignedToId:
          type: string
          format: uuid
        dueDate:
          type: string
          format: date-time
        scheduledTime:
          type: string
          format: time
        recurrenceRule:
          type: object
          properties:
            frequency:
              type: string
              enum: [daily, weekly, monthly]
            interval:
              type: integer
            daysOfWeek:
              type: array
              items:
                type: integer
            dayOfMonth:
              type: integer
        checklistItems:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          $ref: '#/components/schemas/TaskPriority'
        assignedToId:
          type: string
          format: uuid
        dueDate:
          type: string
          format: date-time
    CompleteTaskRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, problem]
        notes:
          type: string
        problemReason:
          $ref: '#/components/schemas/ProblemReason'
        fileKeys:
          type: array
          items:
            type: string
        offlineId:
          type: string
          description: Client-side ID for deduplication
    CompleteChecklistItemRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [ok, problem]
        problemReason:
          $ref: '#/components/schemas/ProblemReason'
        fileKeys:
          type: array
          items:
            type: string
        offlineId:
          type: string
...